feat(v2): make clickable sidebar logo
